refactor: better name for local soil data actions file
